requirements.txt
# 核心框架
gradio>=5.16.0
strands-agents>=0.1.3
strands-agents-tools>=0.1.2

# AI和云服务
boto3>=1.34.0
anthropic>=0.25.0

# 数据处理
pandas>=2.0.0
pydantic>=2.0.0

# 异步和工具
aiofiles
python-dotenv
typing-extensions

# 可观测性（Strands内置）
opentelemetry-api
opentelemetry-sdk

# 其他依赖
numpy
markupsafe
jinja2